Extract common functionality and reduce code duplication
Claude at work
